packages/tools/deps/ussy-stratax/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ussy-stratax"
version = "0.1.0"
description = "Geological Semver for Codebase Archeology — measure behavioral stability of dependencies"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Strata Contributors"}]
keywords = ["semver", "dependencies", "stability", "behavioral-probes", "geological"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "requests>=2.25",
    "pyyaml>=5.4",
    "ussy-core",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov"]

[project.scripts]
ussy-stratax = "ussy_stratax.cli:main"
strata = "ussy_stratax.legacy:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ussy_stratax"]

[tool.uv.sources]
ussy-core = { workspace = true }

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
